Recommend multiple target infrastructure consisting of VMs
* Recommend multiple VM infrastructures for the input csp and region pair
  - Note: The user will have to choose among these.
* Split the files into vm-infra.go and container-infra.go for recommendations
* Add new functions to recommend VM specs and OS images based on source infrastructure
* Add matrix transposition utility for organizing VM recommendations
* Rename variables, keywords, and etc., to reflect the frequency with which they appear in the project
