+++ api/swagger.json
@@ -430,7 +430,7 @@
         },
         "/recommendation/mci": {
             "post": {
-                "description": "Recommend an appropriate multi-cloud infrastructure (MCI) for cloud migration\n\n[Note] `desiredProvider` and `desiredRegion` are required.\n- `desiredProvider` and `desiredRegion` can set on the query parameter or the request body.\n\n- If desiredProvider and desiredRegion are set on request body, the values in the query parameter will be ignored.",
+                "description": "Recommend an appropriate multi-cloud infrastructure (MCI) for cloud migration\n\n[Note] `desiredCsp` and `desiredRegion` are required.\n- `desiredCsp` and `desiredRegion` can set on the query parameter or the request body.\n\n- If desiredCsp and desiredRegion are set on request body, the values in the query parameter will be ignored.",
                 "consumes": [
                     "application/json"
                 ],
@@ -449,7 +449,7 @@
                         "in": "body",
                         "required": true,
                         "schema": {
-                            "$ref": "#/definitions/controller.RecommendInfraRequest"
+                            "$ref": "#/definitions/controller.RecommendVmInfraRequest"
                         }
                     },
                     {
@@ -462,7 +462,7 @@
                         "type": "string",
                         "default": "aws",
                         "description": "Provider (e.g., aws, azure, gcp)",
-                        "name": "desiredProvider",
+                        "name": "desiredCsp",
                         "in": "query"
                     },
                     {
@@ -483,7 +483,7 @@
                     "200": {
                         "description": "The result of recommended infrastructure",
                         "schema": {
-                            "$ref": "#/definitions/controller.RecommendInfraResponse"
+                            "$ref": "#/definitions/controller.RecommendVmInfraResponse"
                         }
                     },
                     "404": {
@@ -917,6 +917,14 @@
                     "type": "string",
                     "example": "mci01"
                 },
+                "postCommand": {
+                    "description": "PostCommand is for the command to bootstrap the VMs",
+                    "allOf": [
+                        {
+                            "$ref": "#/definitions/model.MciCmdReq"
+                        }
+                    ]
+                },
                 "systemLabel": {
                     "description": "SystemLabel is for describing the mci in a keyword (any string can be used) for special System purpose",
                     "type": "string",
@@ -961,6 +969,14 @@
                 "name": {
                     "type": "string",
                     "example": "mci01"
+                },
+                "postCommand": {
+                    "description": "PostCommand is for the command to bootstrap the VMs",
+                    "allOf": [
+                        {
+                            "$ref": "#/definitions/model.MciCmdReq"
+                        }
+                    ]
                 },
                 "systemLabel": {
                     "description": "SystemLabel is for describing the mci in a keyword (any string can be used) for special System purpose",
@@ -1023,6 +1039,34 @@
             }
         },
         "controller.RecommendInfraResponse": {
+            "type": "object",
+            "properties": {
+                "description": {
+                    "type": "string"
+                },
+                "status": {
+                    "type": "string"
+                },
+                "targetInfra": {
+                    "$ref": "#/definitions/model.TbMciDynamicReq"
+                }
+            }
+        },
+        "controller.RecommendVmInfraRequest": {
+            "type": "object",
+            "required": [
+                "onpremiseInfraModel"
+            ],
+            "properties": {
+                "desiredCspAndRegionPair": {
+                    "$ref": "#/definitions/recommendation.CspRegionPair"
+                },
+                "onpremiseInfraModel": {
+                    "$ref": "#/definitions/inframodel.OnpremInfra"
+                }
+            }
+        },
+        "controller.RecommendVmInfraResponse": {
             "type": "object",
             "properties": {
                 "description": {
@@ -1427,6 +1471,27 @@
                 }
             }
         },
+        "model.MciCmdReq": {
+            "type": "object",
+            "required": [
+                "command"
+            ],
+            "properties": {
+                "command": {
+                    "type": "array",
+                    "items": {
+                        "type": "string"
+                    },
+                    "example": [
+                        "client_ip=$(echo $SSH_CLIENT | awk '{print $1}'); echo SSH client IP is: $client_ip"
+                    ]
+                },
+                "userName": {
+                    "type": "string",
+                    "example": "cb-user"
+                }
+            }
+        },
         "model.MciStatusInfo": {
             "type": "object",
             "properties": {
@@ -1564,6 +1629,14 @@
                 "name": {
                     "type": "string",
                     "example": "mci01"
+                },
+                "postCommand": {
+                    "description": "PostCommand is for the command to bootstrap the VMs",
+                    "allOf": [
+                        {
+                            "$ref": "#/definitions/model.MciCmdReq"
+                        }
+                    ]
                 },
                 "systemLabel": {
                     "description": "SystemLabel is for describing the mci in a keyword (any string can be used) for special System purpose",
@@ -1704,6 +1777,19 @@
                     "description": "Uid is universally unique identifier for the object, used for labelSelector",
                     "type": "string",
                     "example": "wef12awefadf1221edcf"
+                }
+            }
+        },
+        "recommendation.CspRegionPair": {
+            "type": "object",
+            "properties": {
+                "csp": {
+                    "type": "string",
+                    "example": "aws"
+                },
+                "region": {
+                    "type": "string",
+                    "example": "ap-northeast-2"
                 }
             }
         }
